+++ clang/docs/ReleaseNotes.rst
@@ -227,6 +227,7 @@ C2y Feature Support
 C23 Feature Support
 ^^^^^^^^^^^^^^^^^^^
 - Clang now allows C23 ``constexpr`` struct member access through the dot operator in constant expressions. (#GH178349)
+- Clang now supports the C23 ``H``, ``D``, and ``DD`` length modifiers. (#GH116962)
 
 Objective-C Language Changes
 -----------------------------